Shift isQuestionSkipped to FeedbackQuestionDetails
