+++ src/main/java/teammates/common/datatransfer/FeedbackResponseDetails.java
@@ -65,20 +65,6 @@ public abstract class FeedbackResponseDetails {
         return getAnswerCsv(question.getQuestionDetails());
     }
     
-    public boolean isQuestionSkipped(Map<String, String[]> requestParameters, int questionIndx, int responseIndx) {
-        String[] answer = HttpRequestHelper.getValuesFromParamMap(requestParameters, Const.ParamsNames.FEEDBACK_RESPONSE_TEXT+"-"+questionIndx+"-"+responseIndx);
-        
-        boolean allAnswersEmpty = true;
-        if(answer!=null){
-            for(int i=0 ; i<answer.length ; i++){
-                if(answer[i]!=null || !answer[i].trim().isEmpty()){
-                    allAnswersEmpty = false;
-                }
-            }
-        }
-        
-        return answer != null && !allAnswersEmpty;
-    }
     
     public static FeedbackResponseDetails createResponseDetails(
             String[] answer, FeedbackQuestionType questionType,
