+++ src/main/java/teammates/ui/controller/FeedbackQuestionSubmissionEditSaveAction.java
@@ -174,12 +174,12 @@ public abstract class FeedbackQuestionSubmissionEditSaveAction extends Action {
         //This field can be null if the question is skipped
         String[] answer = HttpRequestHelper.getValuesFromParamMap(requestParameters, Const.ParamsNames.FEEDBACK_RESPONSE_TEXT+"-"+questionIndx+"-"+responseIndx);
         
-        FeedbackResponseDetails responseDetails = 
-                FeedbackResponseDetails.createResponseDetails(
-                        answer,
-                        questionDetails.questionType,
-                        questionDetails);
-        if (responseDetails.isQuestionSkipped(requestParameters, questionIndx, responseIndx)) {
+        if (!questionDetails.isQuestionSkipped(requestParameters, questionIndx, responseIndx)) {
+            FeedbackResponseDetails responseDetails = 
+                    FeedbackResponseDetails.createResponseDetails(
+                            answer,
+                            questionDetails.questionType,
+                            questionDetails);
             response.setResponseDetails(responseDetails);
         } else {
             response.responseMetaData = new Text("");
